feature(addlogWeekCalendar): creating file for calendar Delegate
